+++ packages/lib/createTestMountParameters/index.ts
@@ -60,6 +60,7 @@ export default (): MockParameters => {
       errors: [],
       hasSmallDisplay: false,
       moveHandle: 0,
+      moveHandleActionButton: 0,
       plugin: {},
       language: '',
       zoomLevel: 0,